# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api",
        "description": "API endpoint",
        "version": "1.0.0"
    },
    "paths": {
        "/api/": {},
        "/api/courses": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Get courses",
                "operationId": "6e1ebb53ff07f5fdf9edd91438a0ec06",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoursesResponse"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "courses"
                ],
                "summary": "Create courses",
                "operationId": "168060ad0097434d27afb2a0b58871ad",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CoursesRequestsBodies"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoursesResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/courses/{courseId}": {
            "get": {
                "tags": [
                    "courses"
                ],
                "summary": "Get course",
                "operationId": "ad1ad293a2549d52caa5b65cbcbdfeff",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "Get course",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoursesResponse"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "courses"
                ],
                "summary": "Update courses",
                "operationId": "376ada19e72f501f9ea57719a43b5103",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "Get course",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CoursesRequestsBodies"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CoursesResponse"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "courses"
                ],
                "summary": "Delete course",
                "operationId": "9742b020fee1602954f5968c687eef62",
                "parameters": [
                    {
                        "name": "courseId",
                        "in": "path",
                        "description": "Delete course",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DeleteCoursesResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lessons": {
            "get": {
                "tags": [
                    "lessons"
                ],
                "summary": "Get lessons",
                "operationId": "aa93ae00e5becacff8d9f8df36f48149",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": [
                                    {
                                        "type": "integer",
                                        "example": "3",
                                        "property": "lessonId"
                                    },
                                    {
                                        "type": "string",
                                        "example": "Zula Ferry",
                                        "property": "name"
                                    },
                                    {
                                        "type": "string",
                                        "example": "Et qui aliquam deleniti non dolorem. Dolore incidunt magni eveniet eius in ut qui. Praesentium impedit ut velit magni nostrum.",
                                        "property": "description"
                                    }
                                ]
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "lessons"
                ],
                "summary": "Create lessons",
                "operationId": "e396dbec3ca387fbcb969ab796a8cf28",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Zula Ferry"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Et qui aliquam deleniti non dolorem. Dolore incidunt magni eveniet eius in ut qui. Praesentium impedit ut velit magni nostrum."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "lessonId": {
                                            "type": "integer",
                                            "example": "3"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Zula Ferry"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Et qui aliquam deleniti non dolorem. Dolore incidunt magni eveniet eius in ut qui. Praesentium impedit ut velit magni nostrum."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/lessons/{lessonId}": {
            "get": {
                "tags": [
                    "lessons"
                ],
                "summary": "Get lesson",
                "operationId": "f6622ec99975e1bb8a7d31e2f2d389b3",
                "parameters": [
                    {
                        "name": "lessonId",
                        "in": "path",
                        "description": "Get lesson",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "lessonId": {
                                            "type": "integer",
                                            "example": "3"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Zula Ferry"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Et qui aliquam deleniti non dolorem. Dolore incidunt magni eveniet eius in ut qui. Praesentium impedit ut velit magni nostrum."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "lessons"
                ],
                "summary": "Update lesson",
                "operationId": "af53ae04d6845fba8955a65fda2abd39",
                "parameters": [
                    {
                        "name": "lessonId",
                        "in": "path",
                        "description": "Update lesson",
                        "required": true
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Zula Ferry"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Et qui aliquam deleniti non dolorem. Dolore incidunt magni eveniet eius in ut qui. Praesentium impedit ut velit magni nostrum."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "lessonId": {
                                            "type": "integer",
                                            "example": "3"
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Zula Ferry"
                                        },
                                        "description": {
                                            "type": "string",
                                            "example": "Et qui aliquam deleniti non dolorem. Dolore incidunt magni eveniet eius in ut qui. Praesentium impedit ut velit magni nostrum."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "lessons"
                ],
                "summary": "Delete lesson",
                "operationId": "794a2c9a457cf203ac5453bdbded6d99",
                "parameters": [
                    {
                        "name": "lessonId",
                        "in": "path",
                        "description": "Delete lesson",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "удалено"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "CoursesRequestsBodies": {
                "required": [
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "PHP"
                    },
                    "description": {
                        "type": "string",
                        "example": "PHP Courses"
                    }
                },
                "type": "object"
            },
            "CoursesResponse": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/CourseSchemas"
                        }
                    }
                },
                "type": "object"
            },
            "DeleteCoursesResponse": {
                "required": [
                    "data"
                ],
                "properties": {
                    "data": {
                        "type": "object",
                        "example": {
                            "message": "удалено"
                        }
                    }
                },
                "type": "object"
            },
            "CourseSchemas": {
                "required": [
                    "courseId",
                    "name",
                    "description"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "maxLength": 16,
                        "example": "abc"
                    },
                    "courseId": {
                        "type": "integer",
                        "maxLength": 64,
                        "example": 2
                    },
                    "description": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "HELLO"
                    }
                },
                "type": "object"
            }
        },
        "requestBodies": {
            "CourseRequestBody": {
                "description": "Course request body",
                "required": true,
                "content": {
                    "application/json": {
                        "schema": {
                            "$ref": "#/components/schemas/CoursesRequestsBodies"
                        }
                    }
                }
            }
        }
    }
}